Tbao thêm lô

+++ WEB_520_Dynamic/Controllers/LoController.cs
@@ -9,6 +9,15 @@ namespace WEB_520_Dynamic.Controllers
 {
 	public class LoController : Controller
 	{
+		public class PrivateResponse
+		{
+			public string type { get; set; }
+			public string res { get; set; }
+
+			//public PrivateResponse() { }
+			public PrivateResponse(string res) { this.res = res; }
+			public PrivateResponse(string type, string res) { this.type = type; this.res = res; }
+		}
 		private readonly ApplicationDbContext _db;
 		public LoController(ApplicationDbContext db)
 		{
@@ -23,7 +32,7 @@ namespace WEB_520_Dynamic.Controllers
 			
 			return View(groupSpL);
 		}
-		public IActionResult ThemLo()
+		/*public IActionResult ThemLo()
 		{
 			IEnumerable<SelectListItem> SP = _db.SAN_PHAMs.Where(x => x.TrangThai == true).Select(
 
@@ -34,7 +43,7 @@ namespace WEB_520_Dynamic.Controllers
 			   });
 			ViewBag.SanPham = SP;
 			return View();
-		}
+		}*/
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public IActionResult ThemLo(LO lo, int id)
@@ -47,23 +56,26 @@ namespace WEB_520_Dynamic.Controllers
 			Console.WriteLine("Mã biên lai : " + id);
 			lo.MaSanPham = lo.SAN_PHAM.MaSanPham;
 			lo.SAN_PHAM = _db.SAN_PHAMs.Where(x => x.MaSanPham == lo.SAN_PHAM.MaSanPham).First();
-			
-			if (ModelState.IsValid)
+            try
+            {
+                foreach (var e in ModelState["lo.Soluong"].Errors)
+                {
+                    if (e.ErrorMessage.Contains("is invalid") || e.ErrorMessage.Contains("is not valid")) { ModelState["lo.SoLuong"].Errors.Remove(e); ModelState["lo.SoLuong"].Errors.Add("Số Lượng Không hợp lệ"); }
+                }
+                foreach (var e in ModelState["lo.TenLo"].Errors)
+                {
+                    if (e.ErrorMessage.Contains("is invalid") || e.ErrorMessage.Contains("is not valid")) { ModelState["lo.TenLo"].Errors.Remove(e); ModelState["lo.TenLo"].Errors.Add("Tên Lô Không hợp lệ"); }
+                }
+            }
+            catch (Exception e)
+            {
+                // ke me t
+            }
+            if (ModelState.IsValid)
 			{
 				_db.LOs.Add(lo);
 				_db.SaveChanges();
-				/*IEnumerable<EntityEntry> es = _db.ChangeTracker.Entries().Where(e => e.State == EntityState.Added);
-				IEnumerable<EntityEntry> es2 = _db.ChangeTracker.Entries().Where(e => e.State == EntityState.Modified);
-				IEnumerable<EntityEntry> es3 = _db.ChangeTracker.Entries().Where(e => e.State == EntityState.Deleted);
-				foreach(EntityEntry e in es)
-				{
-					Console.WriteLine(e.ToString());
-				}
-				foreach(EntityEntry e in es2)
-				{ Console.WriteLine(e.ToString()); }
-				foreach(EntityEntry e in es3)
-				{ Console.WriteLine(e.ToString()); }*/
-
+				
 				//thêm biên lai chi tiết với mã biên lai = 
 				BIEN_LAI_CHI_TIET bienLaiCT = new BIEN_LAI_CHI_TIET();
 				//bienLaiCT.BIEN_LAI = _db.BIEN_LAIs.FirstOrDefault(x => x.MaBienLai == id);
@@ -74,30 +86,56 @@ namespace WEB_520_Dynamic.Controllers
 				_db.BIEN_LAI_CHI_TIETs.Add(bienLaiCT);
 				_db.SaveChanges();
 				TempData["ThongBao"] = "Thêm lô thành công";
-				return RedirectToAction("Index", "BienLaiChiTiet", new { id = bienLaiCT.MaBienLai });
+				//return RedirectToAction("Index", "BienLaiChiTiet", new { id = bienLaiCT.MaBienLai });
+				var j = Json(new PrivateResponse("link", "/BienLaiChiTiet/Index/" + bienLaiCT.MaBienLai));
+				
+				return j;
 				//return RedirectToAction("Index", "BienLaiChiTiet");
 			}
-			return View(lo);
-		}
-		public IActionResult XoaLo(int? ID)
+			string s = "";
+			foreach (var i in ModelState.Values)
+			{
+				if (i.Errors.Count() > 0)
+				{
+					foreach(var j in i.Errors)
+					{
+                        s = s + j.ErrorMessage + "\n";
+                    }
+				}
+			}
+			// return "Thêm lô thất bại" as utf8 encoded string
+			return Json(new PrivateResponse("error", s));
+
+
+        }
+        public IActionResult XoaLo(int? ID)
 		{
 			if (ID == null || ID == 0)
 			{
 				return NotFound();
 			}
-			BIEN_LAI_CHI_TIET bienLaiCT = _db.BIEN_LAI_CHI_TIETs.FirstOrDefault(x => x.MaLo == ID && x.BIEN_LAI.LoaiBienLai == false);
-			LO lo = _db.LOs.FirstOrDefault(x => x.MaLo == ID);
-			if (lo == null)
+			BIEN_LAI_CHI_TIET bienLaiCT = _db.BIEN_LAI_CHI_TIETs.FirstOrDefault(x => x.MaLo == ID && x.BIEN_LAI.LoaiBienLai == false );
+			BIEN_LAI bienlai = _db.BIEN_LAIs.FirstOrDefault(x => x.MaBienLai == bienLaiCT.MaBienLai);
+			if (bienlai.TrangThai == 0)
 			{
-				return NotFound();
+				LO lo = _db.LOs.FirstOrDefault(x => x.MaLo == ID);
+				if (lo == null)
+				{
+					return NotFound();
+				}
+				else
+				{
+					_db.LOs.Remove(lo);
+					_db.SaveChanges();
+					TempData["ThongBaoXoa"] = "Xóa lô thành công";
+				}
+
 			}
 			else
 			{
-				_db.LOs.Remove(lo);
-				_db.SaveChanges();
-				TempData["ThongBaoXoa"] = "Xóa Sàn Phẩm thành công";
+				TempData["ThongBaoXoaLok"] = "Xóa lô không thành công";
 			}
-			
+
 			return RedirectToAction("Index", "BienLaiChiTiet", new { id = bienLaiCT.MaBienLai});
 		}
 		[HttpPost]
@@ -108,12 +146,20 @@ namespace WEB_520_Dynamic.Controllers
 			Console.WriteLine("Mã biên lai : " + id);
 			Console.WriteLine("Mã lô: " + bienLai.MaLo);
 			Console.WriteLine("Số lượng: " + bienLai.SoLuong);
+			try
+			{
+				foreach (var e in ModelState["Soluong"].Errors)
+				{
+					if (e.ErrorMessage.Contains("is invalid") || e.ErrorMessage.Contains("is not valid")) { ModelState["SoLuong"].Errors.Remove(e); ModelState["SoLuong"].Errors.Add("Số Lượng Không hợp lệ"); }
+				}
+			
+			}
+			catch (Exception e)
+			{
+				// ke me t
+			}
 			if (ModelState.IsValid)
 			{
-				
-				//thêm biên lai chi tiết với mã biên lai = 
-				/*BIEN_LAI_CHI_TIET bienLaiCT = new BIEN_LAI_CHI_TIET();
-				bienLaiCT.MaBienLai = id;*/
 				bienLai.MaBienLai = id;
 				Console.WriteLine(bienLai.MaBienLai + " " + bienLai.MaLo + " " + bienLai.SoLuong);
 				_db.BIEN_LAI_CHI_TIETs.Add(bienLai);
@@ -123,8 +169,24 @@ namespace WEB_520_Dynamic.Controllers
 				_db.LOs.Update(lo);
 				_db.SaveChanges();
 				TempData["ThongBao"] = "Thêm lô thành công";
-				return RedirectToAction("BienLaiCTXuat", "BienLaiChiTiet", new { id = bienLai.MaBienLai });
+				//return RedirectToAction("BienLaiCTXuat", "BienLaiChiTiet", new { id = bienLai.MaBienLai });
+				var j = Json(new PrivateResponse("link", "/BienLaiChiTiet/Index/" + bienLai.MaBienLai));
+
+				return j;
 			}
+			string s = "";
+			foreach (var i in ModelState.Values)
+			{
+				if (i.Errors.Count() > 0)
+				{
+					foreach (var j in i.Errors)
+					{
+						s = s + j.ErrorMessage + "\n";
+					}
+				}
+			}
+			// return "Thêm lô thất bại" as utf8 encoded string
+			return Json(new PrivateResponse("error", s));
 			return View(bienLai);
 		}
 		public IActionResult XoaLoXuat(int? ID)
@@ -134,25 +196,36 @@ namespace WEB_520_Dynamic.Controllers
 				return NotFound();
 			}
 			BIEN_LAI_CHI_TIET bienLaiCT = _db.BIEN_LAI_CHI_TIETs.FirstOrDefault(x => x.MaLo == ID && x.BIEN_LAI.LoaiBienLai == true);
+			BIEN_LAI bienlai = _db.BIEN_LAIs.FirstOrDefault(x => x.MaBienLai == bienLaiCT.MaBienLai);
 			Console.WriteLine("Mã biên lai : " + bienLaiCT.MaBienLai);
-			Console.WriteLine("Mã lô: " + bienLaiCT.MaLo);
-			//Console.WriteLine("Loại Biên lai: " + bienLaiCT.BIEN_LAI.LoaiBienLai);
-			LO lo = _db.LOs.FirstOrDefault(x => x.MaLo == ID);
-			if (lo == null)
+			Console.WriteLine( "trạng thái biên lai: "+bienlai.TrangThai);
+			if (bienlai.TrangThai == 0)
 			{
-				return NotFound();
+				LO lo = _db.LOs.FirstOrDefault(x => x.MaLo == ID);
+				if (lo == null)
+				{
+					return NotFound();
+				}
+				else
+				{
+					LO lobandau = _db.LOs.Where(x => x.MaLo == ID).FirstOrDefault();
+					lobandau.SoLuong += bienLaiCT.SoLuong;
+					_db.LOs.Update(lobandau);
+					_db.SaveChanges();
+					_db.BIEN_LAI_CHI_TIETs.Remove(bienLaiCT);
+					_db.SaveChanges();
+
+					TempData["ThongBaoXoak"] = "Xóa Sàn Phẩm thành công";
+				}
 			}
 			else
 			{
-				LO lobandau = _db.LOs.Where(x=> x.MaLo == ID).FirstOrDefault();
-				lobandau.SoLuong += bienLaiCT.SoLuong;
-				_db.LOs.Update(lobandau);
-				_db.SaveChanges();
-				_db.BIEN_LAI_CHI_TIETs.Remove(bienLaiCT);
-				_db.SaveChanges();
-
-				TempData["ThongBaoXoa"] = "Xóa Sàn Phẩm thành công";
+				TempData["ThongBaoXoak"] = "Xóa Sàn Phẩm không thành công";
 			}
+			Console.WriteLine("Mã biên lai : " + bienLaiCT.MaBienLai);
+			Console.WriteLine("Mã lô: " + bienLaiCT.MaLo);
+			//Console.WriteLine("Loại Biên lai: " + bienLaiCT.BIEN_LAI.LoaiBienLai);
+			
 
 			return RedirectToAction("BienLaiCTXuat", "BienLaiChiTiet", new { id = bienLaiCT.MaBienLai });
 		}
